+++ ops/testing.py
@@ -1733,12 +1733,9 @@ class Harness(Generic[CharmType]):
         )
 
     @property
-    def last_rebooted(self) -> Optional[datetime.datetime]:
-        """The time that the charm last called :meth:`ops.Unit.reboot` (with or without *now*).
-
-        Returns ``None`` if :meth:`ops.Unit.reboot` has not been called.
-        """
-        return self._backend._last_reboot
+    def reboot_count(self) -> int:
+        """Count of times that the charm has called :meth:`ops.Unit.reboot`."""
+        return self._backend._reboot_count
 
 
 def _get_app_or_unit_name(app_or_unit: AppUnitOrName) -> str:
@@ -1974,7 +1971,7 @@ class _TestingModelBackend:
         self._secrets: List[_Secret] = []
         self._opened_ports: Set[model.Port] = set()
         self._networks: Dict[Tuple[Optional[str], Optional[int]], _NetworkDict] = {}
-        self._last_reboot = None
+        self._reboot_count = 0
 
     def _validate_relation_access(self, relation_name: str, relations: List[model.Relation]):
         """Ensures that the named relation exists/has been added.
@@ -2536,7 +2533,7 @@ class _TestingModelBackend:
             raise model.ModelError(f'ERROR invalid protocol "{protocol}", expected "tcp", "udp", or "icmp"\n')  # NOQA: test_quote_backslashes
 
     def reboot(self, now: bool = False):
-        self._last_reboot = datetime.datetime.now(datetime.timezone.utc)
+        self._reboot_count += 1
         if not now:
             return
         # This should exit, reboot, and re-emit the event, but we'll need the caller