+++ CHANGES.md
@@ -1,6 +1,6 @@
 # 2.8.0
 
-* Added `Unit.reboot()` and `Harness.last_rebooted``
+* Added `Unit.reboot()` and `Harness.reboot_count``
 * Added `RelationMeta.optional`
 * The type of a `Handle`'s `key` was expanded from `str` to `str|None`
 * Narrowed types of `app` and `unit` in relation events to exclude `None` where applicable